+++ Assets/Tools/ScriptsDataBase/databasetables/Node.cs
@@ -6,11 +6,11 @@ public class Node
 {
 
     [PrimaryKey]
-    public int id { get; set; }
+    public Int64 id { get; set; }
     public string creationDate { get; set; }
 
     public string description { get; set; }
-    public int sectionId { get; set; }
+    public Int64 sectionId { get; set; }
     public string lastUpdate { get; set; }
 
     public override string ToString()
